add missing type hints

+++ backend/salary_calc_fastapi/data/__init__.py
@@ -11,7 +11,7 @@ def datetime_from_string(date: str) -> datetime.date:
     return datetime.strptime(date, "%Y-%m")
 
 
-def _get_value(value_type: str, tax_type: str, category: str, date: Date) -> int:
+def _get_value(value_type: str, tax_type: str, category: str, date: Date) -> float:
 
     values = DATA[value_type][tax_type][category]
 
@@ -26,9 +26,9 @@ def _get_value(value_type: str, tax_type: str, category: str, date: Date) -> int
     return _get_value(value_type, tax_type, category, datetime.now())
 
 
-def get_rate(tax_type: str, category: str, date: Date) -> int:
+def get_rate(tax_type: str, category: str, date: Date) -> float:
     return _get_value("rate", tax_type, category, date)
 
 
-def get_threshold(tax_type: str, category: str, date: Date) -> int:
+def get_threshold(tax_type: str, category: str, date: Date) -> float:
     return _get_value("threshold", tax_type, category, date)
